refactor: split train page in multiple components
