+++ src/pages/train/setDone.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+interface setDoneProps {
+  stats: any;
+}
+
+const SetDone: React.FC<setDoneProps> = ({ ...setDoneProps }) => {
+  const { stats } = setDoneProps;
+
+  return <></>;
+};
+
+export default SetDone;
